Refactored tests
Moved editor related tests to `ForceHexColorTest.ts`. As well as adding force_hex_color specific tests to `StylesTest.ts`
